+++ src/primaite/simulator/system/software.py
@@ -1,4 +1,6 @@
+import copy
 from abc import abstractmethod
+from datetime import datetime
 from enum import Enum
 from ipaddress import IPv4Address
 from typing import Any, Dict, Optional
@@ -231,7 +233,7 @@ class IOSoftware(Software):
 
     installing_count: int = 0
     "The number of times the software has been installed. Default is 0."
-    max_sessions: int = 1
+    max_sessions: int = 100
     "The maximum number of sessions that the software can handle simultaneously. Default is 0."
     tcp: bool = True
     "Indicates if the software uses TCP protocol for communication. Default is True."
@@ -239,6 +241,8 @@ class IOSoftware(Software):
     "Indicates if the software uses UDP protocol for communication. Default is True."
     port: Port
     "The port to which the software is connected."
+    _connections: Dict[str, Dict] = {}
+    "Active connections."
 
     def set_original_state(self):
         """Sets the original state."""
@@ -282,6 +286,65 @@ class IOSoftware(Software):
             _LOGGER.debug(f"{self.name} Error: {self.software_manager.node.hostname} is not online.")
             return False
         return True
+
+    @property
+    def connections(self) -> Dict[str, Dict]:
+        """Return the public version of connections."""
+        return copy.copy(self._connections)
+
+    def add_connection(self, connection_id: str, session_id: Optional[str] = None) -> bool:
+        """
+        Create a new connection to this service.
+
+        Returns true if connection successfully created
+
+        :param: connection_id: UUID of the connection to create
+        :type: string
+        """
+        # if over or at capacity, set to overwhelmed
+        if len(self._connections) >= self.max_sessions:
+            self.health_state_actual = SoftwareHealthState.OVERWHELMED
+            self.sys_log.error(f"{self.name}: Connect request for {connection_id=} declined. Service is at capacity.")
+            return False
+        else:
+            # if service was previously overwhelmed, set to good because there is enough space for connections
+            if self.health_state_actual == SoftwareHealthState.OVERWHELMED:
+                self.health_state_actual = SoftwareHealthState.GOOD
+
+            # check that connection already doesn't exist
+            if not self._connections.get(connection_id):
+                session_details = None
+                if session_id:
+                    session_details = self._get_session_details(session_id)
+                self._connections[connection_id] = {
+                    "ip_address": session_details.with_ip_address if session_details else None,
+                    "time": datetime.now(),
+                }
+                self.sys_log.info(f"{self.name}: Connect request for {connection_id=} authorised")
+                return True
+            # connection with given id already exists
+            self.sys_log.error(
+                f"{self.name}: Connect request for {connection_id=} declined. Connection already exists."
+            )
+            return False
+
+    def remove_connection(self, connection_id: str) -> bool:
+        """
+        Remove a connection from this service.
+
+        Returns true if connection successfully removed
+
+        :param: connection_id: UUID of the connection to create
+        :type: string
+        """
+        if self.connections.get(connection_id):
+            self._connections.pop(connection_id)
+            self.sys_log.info(f"{self.name}: Connection {connection_id=} closed.")
+            return True
+
+    def clear_connections(self):
+        """Clears all the connections from the software."""
+        self._connections = {}
 
     def send(
         self,