+++ tests/unit_tests/_primaite/_simulator/_system/_services/test_services.py
@@ -1,3 +1,5 @@
+from uuid import uuid4
+
 from primaite.simulator.system.services.service import ServiceOperatingState
 from primaite.simulator.system.software import SoftwareHealthState
 
@@ -66,3 +68,34 @@ def test_enable_disable(service):
 
     service.enable()
     assert service.operating_state == ServiceOperatingState.STOPPED
+
+
+def test_overwhelm_service(service):
+    service.max_sessions = 2
+    service.start()
+
+    uuid = str(uuid4())
+    assert service.add_connection(connection_id=uuid)  # should be true
+    assert service.health_state_actual is SoftwareHealthState.GOOD
+
+    assert not service.add_connection(connection_id=uuid)  # fails because connection already exists
+    assert service.health_state_actual is SoftwareHealthState.GOOD
+
+    assert service.add_connection(connection_id=str(uuid4()))  # succeed
+    assert service.health_state_actual is SoftwareHealthState.GOOD
+
+    assert not service.add_connection(connection_id=str(uuid4()))  # fail because at capacity
+    assert service.health_state_actual is SoftwareHealthState.OVERWHELMED
+
+
+def test_create_and_remove_connections(service):
+    service.start()
+    uuid = str(uuid4())
+
+    assert service.add_connection(connection_id=uuid)  # should be true
+    assert len(service.connections) == 1
+    assert service.health_state_actual is SoftwareHealthState.GOOD
+
+    assert service.remove_connection(connection_id=uuid)  # should be true
+    assert len(service.connections) == 0
+    assert service.health_state_actual is SoftwareHealthState.GOOD